Fix: use only one add new language button
