+++ frontend/src/components/pages/UserProfilePage.tsx
@@ -13,9 +13,6 @@ import {
 import {
   MANAGE_USERS_TABLE_DELETE_USER_BUTTON,
   MANAGE_USERS_TABLE_DELETE_USER_CONFIRMATION,
-  NEW_BOOK_TEST_ADDED_BUTTON,
-  NEW_BOOK_TEST_ADDED_HEADING,
-  NEW_BOOK_TEST_ADDED_MESSAGE,
 } from "../../utils/Copy";
 import ConfirmationModal from "../utils/ConfirmationModal";
 import { GET_USER, User } from "../../APIClients/queries/UserQueries";
@@ -27,15 +24,10 @@ import {
   ApprovedLanguagesMap,
   parseApprovedLanguages,
 } from "../../utils/Utils";
-import ApprovedLanguagesTable from "../admin/ApprovedLanguagesTable";
+import ApprovedLanguagesTable from "../userProfile/ApprovedLanguagesTable";
 import AssignedStoryTranslationsTable from "../userProfile/AssignedStoryTranslationsTable";
-import NewLanguageModal from "./NewLanguageModal";
 import { StoryAssignStage } from "../../constants/Enums";
 import InfoAlert from "../utils/InfoAlert";
-import {
-  CREATE_TRANSLATION_TEST,
-  CreateTranslationTestResponse,
-} from "../../APIClients/mutations/StoryMutations";
 import { useFileDownload } from "../../utils/FileUtils";
 
 type UserProfilePageProps = {
@@ -70,9 +62,6 @@ const UserProfilePage = () => {
   const [storyAssignStage, setStoryAssignStage] = useState<StoryAssignStage>(
     StoryAssignStage.INITIAL,
   );
-
-  const [addNewLanguage, setAddNewLanguage] = useState(false);
-  const [alertNewTest, setAlertNewTest] = useState(false);
 
   const history = useHistory();
 
@@ -132,40 +121,6 @@ const UserProfilePage = () => {
 
   const filterStyle = useStyleConfig("Filter");
 
-  const [createTranslationTest] = useMutation<{
-    response: CreateTranslationTestResponse;
-  }>(CREATE_TRANSLATION_TEST);
-
-  const openNewLanguageModal = () => {
-    setAddNewLanguage(true);
-  };
-
-  const closeNewLanguageModal = () => {
-    setAddNewLanguage(false);
-  };
-
-  const closeNewTestModal = () => {
-    setAlertNewTest(false);
-  };
-
-  const addNewTest = async (newLanguage: string) => {
-    try {
-      await createTranslationTest({
-        variables: {
-          userId,
-          level: 2,
-          language: newLanguage,
-        },
-      });
-      closeNewLanguageModal();
-      setAlertNewTest(true);
-    } catch (err) {
-      window.alert(err);
-      closeNewLanguageModal();
-      closeNewTestModal();
-    }
-  };
-
   if (loading) return <div />;
   if (error) return <Redirect to="/404" />;
 
@@ -203,14 +158,6 @@ const UserProfilePage = () => {
           {!isAdmin && <UserProfileForm isSignup={false} />}
           <Flex justifyContent="space-between" margin="40px 0 10px 0">
             <Heading size="lg"> Approved Languages & Levels </Heading>
-            <Button
-              colorScheme="blue"
-              onClick={() => openNewLanguageModal()}
-              variant="blueOutline"
-              width="250px"
-            >
-              ADD NEW LANGUAGE +
-            </Button>
           </Flex>
           {isAdmin &&
             (alert ? (
@@ -285,22 +232,6 @@ const UserProfilePage = () => {
           buttonMessage={MANAGE_USERS_TABLE_DELETE_USER_BUTTON}
         />
       )}
-      {addNewLanguage && (
-        <NewLanguageModal
-          addNewTest={addNewTest}
-          approvedLanguagesTranslation={approvedLanguagesTranslation}
-          isOpen={addNewLanguage}
-          onClose={closeNewLanguageModal}
-        />
-      )}
-      <ConfirmationModal
-        confirmation={alertNewTest}
-        onClose={closeNewTestModal}
-        onConfirmationClick={() => history.push("/")}
-        confirmationHeading={NEW_BOOK_TEST_ADDED_HEADING}
-        confirmationMessage={NEW_BOOK_TEST_ADDED_MESSAGE}
-        buttonMessage={NEW_BOOK_TEST_ADDED_BUTTON}
-      />
     </Flex>
   );
 };
