+++ frontend/src/components/userProfile/AssignStoryModal.tsx
@@ -1,4 +1,5 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
+import { useQuery } from "@apollo/client";
 import Select from "react-select";
 import {
   Button,
@@ -17,56 +18,110 @@ import {
   useStyleConfig,
 } from "@chakra-ui/react";
 
-import { languageOptions } from "../../constants/Languages";
-import { levelOptions } from "../../constants/Levels";
 import { roleOptions } from "../../constants/Roles";
 
+import { buildAssignStoryQuery } from "../../APIClients/queries/StoryQueries";
 import { convertLanguageTitleCase } from "../../utils/LanguageUtils";
-import { ApprovedLanguagesMap } from "../../utils/Utils";
-import DropdownIndicator from "./DropdownIndicator";
+import { ApprovedLanguagesMap, isObjEmpty } from "../../utils/Utils";
+import DropdownIndicator from "../utils/DropdownIndicator";
 import { colourStyles } from "../../theme/components/Select";
 
-export type NewApprovedLanguage = {
-  language: string;
+export type StoryToAssign = {
+  storyId: number;
+  storyTranslationId?: number;
+  title: string;
+  description: string;
   level: number;
-  role: string;
+  language: string;
 };
 
-export type ApprovedLanguageModalProps = {
+export type AssignStoryModalProps = {
+  userId: number;
   isOpen: boolean;
   onClose: () => void;
-  onAssign: (newApprovedLanguage: NewApprovedLanguage) => void;
+  onAssignStory: (story: StoryToAssign) => void;
   approvedLanguagesTranslation: ApprovedLanguagesMap;
   approvedLanguagesReview: ApprovedLanguagesMap;
 };
 
-const ApproveLanguageModal = ({
+const AssignStoryModal = ({
+  userId,
   isOpen,
   onClose,
-  onAssign,
+  onAssignStory,
   approvedLanguagesTranslation,
   approvedLanguagesReview,
-}: ApprovedLanguageModalProps) => {
-  const [role, setRole] = useState<string | null>(null);
+}: AssignStoryModalProps) => {
+  const isApprovedReviewer = !isObjEmpty(approvedLanguagesReview);
+
+  // default selection to Translator if user is not an approved reviewer
+  const [role, setRole] = useState<string | null>(
+    isApprovedReviewer ? null : "Translator",
+  );
   const [language, setLanguage] = useState<string | null>(null);
   const [level, setLevel] = useState<number | null>(null);
-
-  const currentApprovedLanguages = new Set(
-    Object.keys(
-      role === "Translator"
-        ? approvedLanguagesTranslation
-        : approvedLanguagesReview,
-    ),
+  const [stories, setStories] = useState<StoryToAssign[]>([]);
+  const [selectedStory, setSelectedStory] = useState<StoryToAssign | null>(
+    null,
   );
 
-  // don't show languages that the user is already approved for
-  const newLanguageOptions = languageOptions.filter(
-    (lang) => !currentApprovedLanguages.has(lang.value),
+  const query = buildAssignStoryQuery(
+    role === "Translator",
+    language!,
+    level!,
+    userId,
   );
+
+  useQuery(query.string, {
+    fetchPolicy: "cache-and-network",
+    skip: language == null || level == null,
+    onCompleted: (data) => {
+      const storyData = data[query!!.fieldName].map((storyObj: any) => ({
+        ...storyObj,
+        language,
+      }));
+      setStories(storyData);
+    },
+  });
+
+  useEffect(() => {
+    // reset story selection when one of the fields change
+    setSelectedStory(null);
+    setStories([]);
+  }, [language, level, role]);
+
+  useEffect(() => {
+    // reset fields every time role changes
+    setLanguage(null);
+    setLevel(null);
+  }, [role]);
+
+  const approvedLanguages =
+    role === "Translator"
+      ? approvedLanguagesTranslation
+      : approvedLanguagesReview;
+
+  const languageOptions = role
+    ? Object.keys(approvedLanguages).map((value) => ({
+        value,
+      }))
+    : [];
+
+  const maxApprovedLevel = approvedLanguages[language ?? ""] || 0;
+
+  // display levels that the user is approved for
+  const levelOptions = language
+    ? [...Array(maxApprovedLevel + 1).keys()]
+        .slice(1)
+        .map((value) => ({ value: `${value}` }))
+    : [];
 
   const disabledStyle = useStyleConfig("Disabled");
+  const isRoleSelectDisabled = !isApprovedReviewer;
   const isLanguageSelectDisabled = role == null;
-  const isLevelSelectDisabled = role == null;
+  const isLevelSelectDisabled = language == null;
+  const isTitleSelectDisabled =
+    (language == null || level == null) && stories.length === 0;
   return (
     <Modal
       isOpen={isOpen}
@@ -84,10 +139,10 @@ const ApproveLanguageModal = ({
         />
         <ModalHeader paddingTop="20px">
           <Heading as="h3" size="lg">
-            Approve New Language/Level
+            Assign Story
           </Heading>
           <Text fontSize="16px" fontWeight="normal">
-            Please assign the user a new language and level.
+            Please assign an existing story to the user.
           </Text>
         </ModalHeader>
         <ModalBody paddingTop="24px" paddingBottom="36px">
@@ -103,7 +158,10 @@ const ApproveLanguageModal = ({
               <Text marginBottom="12px">
                 Please select a role to assign to the user.
               </Text>
-              <Box width="80%">
+              <Box
+                width="80%"
+                sx={isRoleSelectDisabled ? disabledStyle : undefined}
+              >
                 <Select
                   placeholder="Select role"
                   options={roleOptions}
@@ -112,6 +170,7 @@ const ApproveLanguageModal = ({
                   value={role ? { value: role } : null}
                   styles={colourStyles}
                   components={{ DropdownIndicator }}
+                  isDisabled={isRoleSelectDisabled}
                 />
               </Box>
             </GridItem>
@@ -126,7 +185,7 @@ const ApproveLanguageModal = ({
               >
                 <Select
                   placeholder="Select language"
-                  options={newLanguageOptions}
+                  options={languageOptions}
                   onChange={(option: any) => setLanguage(option.value)}
                   getOptionLabel={(option: any) =>
                     convertLanguageTitleCase(option.value)
@@ -138,7 +197,7 @@ const ApproveLanguageModal = ({
                 />
               </Box>
             </GridItem>
-            <GridItem rowStart={3}>
+            <GridItem rowStart={2}>
               <Heading size="sm">Level</Heading>
               <Text marginBottom="12px">
                 Please select a level to assign to the user.
@@ -161,22 +220,35 @@ const ApproveLanguageModal = ({
                 />
               </Box>
             </GridItem>
+            <GridItem rowStart={3}>
+              <Heading size="sm">Title</Heading>
+              <Text marginBottom="12px">
+                Please select a book title to assign to the user.
+              </Text>
+              <Box
+                width="80%"
+                sx={isTitleSelectDisabled ? disabledStyle : undefined}
+              >
+                <Select
+                  placeholder="Select book title"
+                  options={stories}
+                  onChange={(option: any) => setSelectedStory(option)}
+                  getOptionLabel={(option: any) => option.title}
+                  value={selectedStory}
+                  styles={colourStyles}
+                  components={{ DropdownIndicator }}
+                />
+              </Box>
+            </GridItem>
           </Grid>
           <Flex marginRight="28px" justifyContent="flex-end">
             <Button
               fontSize="14px"
               colorScheme="blue"
-              width="120px"
-              onClick={() =>
-                onAssign({
-                  language: language!,
-                  level: level!,
-                  role: role!,
-                })
-              }
-              isDisabled={!language || !level}
+              onClick={() => onAssignStory(selectedStory!)}
+              isDisabled={selectedStory == null}
             >
-              Assign
+              Assign Story
             </Button>
           </Flex>
         </ModalBody>
@@ -185,4 +257,4 @@ const ApproveLanguageModal = ({
   );
 };
 
-export default ApproveLanguageModal;
+export default AssignStoryModal;
